FIX Sub & Test UX and Fix Login Bug
I have modified the subscription and testimonials section so that whenever a user is logged in, they don't need to type in their name again. After adding that feature, there's a bug on the UI, whenever the user login, the UI doesn't change and I have been working on that for some quite a time. Turns out, the problem is in main.go, I needed to change the config so that it allows Authorization because normally it doesn't.